asoc: aic326x codec: Fix compilation warnigs
Add typecast in switch statement

Bug 1274537

Reviewed-on: http://git-master/r/237042
(cherry picked from commit 3a4195207f762d3ae7720d71a6e8d1f1fc9fe165)
Reviewed-on: http://git-master/r/250768

+++ sound/soc/codecs/aic3xxx/aic3xxx_cfw_ops.c
@@ -430,7 +430,7 @@ static void aic3xxx_cfw_dlcmds(struct cfw_state *ps, struct cfw_block *pb)
 		if (c->cid != CFW_CMD_BRANCH_IM &&
 		    c->cid != CFW_CMD_BRANCH_ID && c->cid != CFW_CMD_NOP)
 			cond = 0;
-		switch (c->cid) {
+		switch ((int)(c->cid)) {
 		case 0 ... (CFW_CMD_NOP - 1):
 			ps->ops->reg_write(ps->codec, c->reg.bpod,
 					   c->reg.data);
